Move deps from tests to root to prevent recursive tests (#59)
When Eventgen and CIM was placed in the tests/deps folder this caused additional tests to be generated by pytest which was increasing run time. Additional I discovered the docker_compose_files is not honored unless it is in the conftest.py in the tests folder so moved that to allow all tests to work.