fix: put helmut's REST connector back into PSA
The reason for it is that some of the modinput tests are using some
of the unpredicted paths of PSA, REST connector in this case.

This is a temporary fix and will be removed as soon as we have time to
refactor modinput tests in some of the repositories (around 12 atm).