fix: add Splunk HEC token validation for external Splunk (#877)
This PR adds validation of HEC token. It just checks if an event can be
send via given HEC token to external Splunk.
Fix was tested locally. The result for 3 workers are presented in the
attached screenshot (with pytest-splunk-addon code from develop branch
it breaks main worker , restarts it and wait indefinitely).
Although splunk_external fixture is defined in the session scope it runs
on each worker separately.
The following question arises: is it expected behaviour, or it's
something we'd like to change (for the fixture logic to be executed
before pytest splits for several processes). I do not believe that we
have to check if Splunk is responsive from each worker separately. That
would need to be explored (such solution might not be compatible with
pytest):

![image](https://github.com/user-attachments/assets/95bd1c7c-d6ed-43d7-8249-7475e4760808)